# perf: add lru_cache to improve performance with multiple themes