These changes should improve the performance when there are plenty of themes defined in **COMPREHENSIVE_THEME_DIRS**, using lru_cache to reduce the number of calls to the method **is_theme_dir**.

We have a bad response time due to the file I/O when it's **running in docker** the more themes you add to your platform the more slow answer you will have.

We can evidence this in these tests:

Test without comprehensive theming (0 Themes)
```bash
running (10m09.6s), 00/28 VUs, 602 complete and 0 interrupted iterations
http_req_duration..............: avg=553.57ms min=107.08ms med=492.97ms max=3.98s p(90)=1.02s p(95)=1.24s
```

Test with indigo (1 Theme)
```bash
running (10m15.3s), 00/28 VUs, 523 complete and 0 interrupted iterations
http_req_duration..............: avg=1.24s min=132.26ms med=1.23s max=2.95s p(90)=2s p(95)=2.17s
```

Test with multiple themes (5 Themes)
```bash
running (10m12.7s), 00/28 VUs, 313 complete and 3 interrupted iterations
http_req_duration..............: avg=4.56s   min=111.86ms med=5.25s   max=7.57s    p(90)=6.25s    p(95)=6.5s
```

**Finally, we can see the improvement with the changes in this PR (5 Themes)**
```bash
running (10m06.6s), 00/28 VUs, 592 complete and 0 interrupted iterations
http_req_duration..............: avg=633.36ms min=101.13ms med=606.9ms  max=2.15s    p(90)=1.13s    p(95)=1.27s 
```

For a thorough description of the issue, performance tests & solution design, you could refer to [
ISSUE: Comprehensive theming performance](https://discuss.openedx.org/t/issue-comprehensive-theming-performance/8292)

---
We may wish to backport all of these changes to Nutmeg