@ormsbee 2022-10-04
FYI @openedx/perf-interest: This is a great perf win, but the bigger takeaway here is that the switch to docker-based deployment looks like it's causing disk I/O to become much more expensive, meaning that we could see performance regressions in a bunch of places that we haven't touched in a long time.

But enough sober talk. Look at the sweet numbers!

Before:

```
avg=3.17s
min=133.3ms
med=3.56s
max=7.38s
p(90)=4.54s
p(95)=4.8s
```

After:

```
avg=633.36ms
min=101.13ms
med=606.9ms 
max=2.15s
p(90)=1.13s
p(95)=1.27s 
```