@ormsbee 2022-10-05
BTW, I think the tests should be able to clear the LRU cache with calls like `get_themes.clear_cache()` (the `@lru_cache` decorator adds this to functions it wraps).