@kdmccormick 2022-10-05
GHA is reporting an error: https://github.com/openedx/edx-platform/actions/runs/3190974009

It says `You have an error in your yaml syntax on line 65`

I think this is an error on master, I'll take a look.