@kdmccormick 2022-10-05
Yup, the error is from master. Once [this fix](https://github.com/openedx/edx-platform/pull/31105) merges, you can rebase and then tests should run.