@mariajgrimaldi 2022-10-06
@ormsbee: not sure either. But each time a first-time contributor opens a PR, someone with permissions needs to approve the workflow each time a new commit is pushed. Not sure what the conditions are for this to happen, but it has happened to me a couple of times.