@openedx-webhooks 2022-10-06
@Alec4r 🎉 Your pull request was merged!  Please take a moment  to [answer a two question survey](https://docs.google.com/forms/d/e/1FAIpQLSceJOyGJ6JOzfy6lyR3T7EW_71OWUnNQXp68Fymsk3MkNoSDg/viewform?usp=pp_url&entry.1671973413=openedx/edx-platform&entry.867055334=https://github.com/openedx/edx-platform/pull/31090&entry.1484655318=https://github.com/Alec4r&entry.752974735=2022-10-04+15:28&entry.1917517419=2022-10-06+18:32&entry.2133058324=Yes) so we can improve your experience in the future.

<!-- comment:end_survey -->