@ormsbee 2022-10-19
@Alec4r: Please add a note about this to the [Olive release notes page](https://openedx.atlassian.net/wiki/spaces/COMM/pages/3396501578/Olive) under the [Operational](https://openedx.atlassian.net/wiki/spaces/COMM/pages/3396501578/Olive#Operational) section. Anybody running a Docker-based deployment with theming will likely be interested in this.